move TraceLogMessages to defered_test
