+++ cmd/srcd/cmd/defered_test.go
@@ -2,11 +2,13 @@ package cmd
 
 import (
 	"bytes"
+	"encoding/json"
+	"io"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/sirupsen/logrus"
-	"github.com/src-d/engine/cmdtests"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
@@ -64,7 +66,7 @@ func (s *DeferedTestSuite) TestPrint() {
 		var memLog bytes.Buffer
 		d := s.buildDefered(false, nil)
 
-		logMessages := cmdtests.TraceLogMessages(s.buildFn(d, 500*time.Millisecond), &memLog)
+		logMessages := TraceLogMessages(s.buildFn(d, 500*time.Millisecond), &memLog)
 
 		require.Equal(len(logMessages), 3)
 		expected := [3]string{"Start", "Hello World!", "End"}
@@ -79,7 +81,7 @@ func (s *DeferedTestSuite) TestPrint() {
 		var memLog bytes.Buffer
 		d := s.buildDefered(false, nil)
 
-		logMessages := cmdtests.TraceLogMessages(s.buildFn(d, 100*time.Millisecond), &memLog)
+		logMessages := TraceLogMessages(s.buildFn(d, 100*time.Millisecond), &memLog)
 
 		require.Equal(len(logMessages), 2)
 		expected := [2]string{"Start", "End"}
@@ -96,7 +98,7 @@ func (s *DeferedTestSuite) TestPrintWithSpinner() {
 		var memLog bytes.Buffer
 		d := s.buildDefered(true, nil)
 
-		logMessages := cmdtests.TraceLogMessages(s.buildFn(d, 500*time.Millisecond), &memLog)
+		logMessages := TraceLogMessages(s.buildFn(d, 500*time.Millisecond), &memLog)
 
 		require.Equal(len(logMessages), 6)
 		expected := [6]string{
@@ -118,7 +120,7 @@ func (s *DeferedTestSuite) TestPrintWithSpinner() {
 		var memLog bytes.Buffer
 		d := s.buildDefered(true, nil)
 
-		logMessages := cmdtests.TraceLogMessages(s.buildFn(d, 100*time.Millisecond), &memLog)
+		logMessages := TraceLogMessages(s.buildFn(d, 100*time.Millisecond), &memLog)
 
 		require.Equal(len(logMessages), 2)
 		expected := [2]string{"Start", "End"}
@@ -151,7 +153,7 @@ func (s *DeferedTestSuite) TestPrintWithInputFn() {
 		var memLog bytes.Buffer
 		d := s.buildDefered(false, inputFn)
 
-		logMessages := cmdtests.TraceLogMessages(s.buildFn(d, 500*time.Millisecond), &memLog)
+		logMessages := TraceLogMessages(s.buildFn(d, 500*time.Millisecond), &memLog)
 
 		require.Equal(len(logMessages), 5)
 		expected := [5]string{
@@ -172,7 +174,7 @@ func (s *DeferedTestSuite) TestPrintWithInputFn() {
 		var memLog bytes.Buffer
 		d := s.buildDefered(false, inputFn)
 
-		logMessages := cmdtests.TraceLogMessages(s.buildFn(d, 100*time.Millisecond), &memLog)
+		logMessages := TraceLogMessages(s.buildFn(d, 100*time.Millisecond), &memLog)
 
 		require.Equal(len(logMessages), 2)
 		expected := [2]string{"Start", "End"}
@@ -180,4 +182,39 @@ func (s *DeferedTestSuite) TestPrintWithInputFn() {
 			require.Equal(lm.Msg, expected[i])
 		}
 	})
+}
+
+type LogMessage struct {
+	Msg   string
+	Time  string
+	Level string
+}
+
+func TraceLogMessages(fn func(), memLog *bytes.Buffer) []LogMessage {
+	logrus.SetOutput(memLog)
+	logrus.SetFormatter(&logrus.JSONFormatter{})
+
+	fn()
+
+	var result []LogMessage
+	if memLog.Len() == 0 {
+		return result
+	}
+
+	dec := json.NewDecoder(strings.NewReader(memLog.String()))
+	for {
+		var i LogMessage
+		err := dec.Decode(&i)
+		if err == io.EOF {
+			break
+		}
+
+		if err != nil {
+			panic(err)
+		}
+
+		result = append(result, i)
+	}
+
+	return result
 }
